Proto1: Tile selector update
Pulled Tile Selector out of the Tile object and put it into the Map
object. Now, when you click a tile on the map, it becomes selected. If
you click a different tile, that tile becomes selected instead.
Note: I left the selector code commented out in the Tile class, in case
this was to be used for a highlighter (in place of cyan color)