Movement over time with basic pathing (Prototype 1)
Added states to characterbehavior. Made it possible for character to
follow a path in the "move" state. These added functions might need to
be moved to another class -- it's gotten messy.

Added some randomly placed collideable blocks. To do this, had to alter
Tile and Map to use "Awake()" instead of "Start()"